+++ src/main/java/com/xtu/plugin/common/OnResultListener.java
@@ -1,8 +1,0 @@
-package com.xtu.plugin.common;
-
-import org.jetbrains.annotations.NotNull;
-
-public interface OnResultListener {
-
-    void onResult(@NotNull String htmlContent);
-}