+++ src/main/java/com/xtu/plugin/previewer/lottie/generator/LottieHtmlGenerator.java
@@ -1,49 +1,41 @@
 package com.xtu.plugin.previewer.lottie.generator;
 
-import com.intellij.openapi.application.ApplicationManager;
-import com.intellij.openapi.application.ReadAction;
+import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vfs.VfsUtil;
 import com.intellij.openapi.vfs.VirtualFile;
-import com.intellij.util.ui.UIUtil;
-import com.xtu.plugin.common.utils.DisplayUtils;
+import com.intellij.ui.JBColor;
+import com.xtu.plugin.common.utils.ColorUtils;
 import com.xtu.plugin.common.utils.FileUtils;
-import com.xtu.plugin.common.OnResultListener;
-import kotlin.Pair;
+import com.xtu.plugin.previewer.HtmlGenerator;
 import org.jetbrains.annotations.NotNull;
-import org.json.JSONObject;
+import org.jetbrains.annotations.Nullable;
 
 
 //Lottie Html生成器
-public class LottieHtmlGenerator {
+public class LottieHtmlGenerator extends HtmlGenerator {
 
-    public static void generate(@NotNull VirtualFile lottieFile,
-                                @NotNull OnResultListener listener) {
-        ReadAction.run(() -> readAndGenerate(lottieFile, listener));
+    public LottieHtmlGenerator(@NotNull Project project,
+                               @NotNull VirtualFile file,
+                               @NotNull HtmlGenerator.OnHTMLGenerateListener readyListener) {
+        super(project, file, readyListener);
     }
 
-    private static void readAndGenerate(@NotNull VirtualFile lottieFile,
-                                        @NotNull OnResultListener listener) {
+    @Override
+    public @Nullable String createHtml() {
         try {
-            String lottieContent = VfsUtil.loadText(lottieFile);
-            JSONObject lottieJson = new JSONObject(lottieContent);
-            int width = (int) lottieJson.optDouble("w", 0);
-            int height = (int) lottieJson.optDouble("h", 0);
-            Pair<Integer, Integer> fitSize = DisplayUtils.getFitSize(width, height);
-            String htmlContent = FileUtils.readTextFromResource("html/lottie/index.html");
-            String cssContent = FileUtils.readTextFromResource("html/lottie/index.css");
-            String lottieJsContent = FileUtils.readTextFromResource("html/lottie/libs/lottie_light.min.js");
-            String jsContent = FileUtils.readTextFromResource("html/lottie/index.js");
-            //拼装html内容
-            String result = htmlContent.replace("{style_placeholder}", cssContent)
+            String lottieContent = VfsUtil.loadText(getFile());
+            String htmlContent = FileUtils.loadTextFromResource("html/lottie/index.html");
+            String cssContent = FileUtils.loadTextFromResource("html/lottie/index.css");
+            String lottieJsContent = FileUtils.loadTextFromResource("html/lottie/libs/lottie_light.min.js");
+            String jsContent = FileUtils.loadTextFromResource("html/lottie/index.js");
+
+            return htmlContent.replace("{style_placeholder}", cssContent)
                     .replace("{script_placeholder}", lottieJsContent + " " + jsContent)
-                    .replace("{body_color}", UIUtil.isUnderDarcula() ? "black" : "white")
-                    .replace("{main_color}", UIUtil.isUnderDarcula() ? "white" : "black")
-                    .replace("{lottie_width}", String.valueOf(fitSize.getFirst() + 20))
-                    .replace("{lottie_height}", String.valueOf(fitSize.getSecond() + 20))
-                    .replace("{data_placeholder}", lottieContent);
-            ApplicationManager.getApplication().invokeLater(() -> listener.onResult(result));
+                    .replace("{body_color}", ColorUtils.toString(JBColor.background()))
+                    .replace("{main_color}", ColorUtils.toString(JBColor.foreground()))
+                    .replace("{file_content_placeholder}", lottieContent);
         } catch (Exception e) {
-            e.printStackTrace();
+            return null;
         }
     }
 }