+++ src/main/java/com/xtu/plugin/previewer/svg/SVGImagePanel.java
@@ -3,109 +3,134 @@ package com.xtu.plugin.previewer.svg;
 import com.intellij.openapi.application.Application;
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.ui.Messages;
+import com.intellij.openapi.util.text.Formats;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.ui.components.JBBox;
+import com.intellij.util.ui.JBUI;
 import com.intellij.util.ui.StartupUiUtil;
+import com.twelvemonkeys.imageio.plugins.svg.SVGImageReader;
+import com.twelvemonkeys.imageio.plugins.svg.SVGReadParam;
+import com.xtu.plugin.common.ui.ScalePanel;
 import com.xtu.plugin.common.utils.ImageUtils;
-import com.xtu.plugin.common.utils.LogUtils;
 import org.jetbrains.annotations.NotNull;
 
-import javax.imageio.ImageReader;
 import javax.swing.*;
 import java.awt.*;
-import java.awt.event.InputEvent;
-import java.awt.event.KeyEvent;
 import java.awt.image.BufferedImage;
 
-public class SVGImagePanel extends JPanel {
+public class SVGImagePanel extends ScalePanel {
 
-    private float curScale = 1.0f;
-    private BufferedImage bufferedImage;
+    private static final int sDefaultSize = 100;
 
-    public SVGImagePanel(@NotNull VirtualFile file, @NotNull String fileExtension) {
+    private final VirtualFile svgFile;
+
+    private BufferedImage image;
+    private JLabel infoLabel;
+    private JLabel errorLabel;
+
+    public SVGImagePanel(@NotNull VirtualFile svgFile) {
+        super();
+        this.svgFile = svgFile;
         this.setLayout(new BorderLayout());
-        this.registerKeyBoardEvent();
-        this.render(file, fileExtension);
+        this.render();
     }
 
-    private void render(@NotNull VirtualFile file, @NotNull String fileExtension) {
+    private void showImageInfo(int width, int height, @NotNull BufferedImage image) {
+        Application application = ApplicationManager.getApplication();
+        application.invokeLater(() -> {
+            if (this.infoLabel == null) {
+                String info = String.format("%dx%d SVG (%d-bit color) %s",
+                        width, height,
+                        image.getColorModel().getPixelSize(),
+                        Formats.formatFileSize(svgFile.getLength())
+                );
+                JBBox container = JBBox.createHorizontalBox();
+                container.setBorder(JBUI.Borders.empty(10));
+                container.add(JBBox.createHorizontalGlue());
+                container.add(this.infoLabel = new JLabel(info));
+                add(container, BorderLayout.NORTH);
+            }
+            if (this.errorLabel != null && this.errorLabel.getParent() != null) {
+                remove(this.errorLabel);
+            }
+            this.image = image;
+            repaint();
+        });
+    }
+
+    private void showErrorInfo() {
+        Application application = ApplicationManager.getApplication();
+        application.invokeLater(() -> {
+            if (this.errorLabel == null) {
+                String errorText = "Fail to load SVG Image";
+                this.errorLabel = new JLabel(errorText, Messages.getErrorIcon(), SwingConstants.CENTER);
+            }
+            if (this.errorLabel.getParent() == null) {
+                add(this.errorLabel, BorderLayout.CENTER);
+            }
+            this.image = null;
+            repaint();
+        });
+    }
+
+    @Override
+    protected void onScaleChanged() {
+        this.render();
+    }
+
+    private void render() {
         Application application = ApplicationManager.getApplication();
         application.executeOnPooledThread(() -> {
-            ImageReader imageReader = ImageUtils.getTwelveMonkeysRead(fileExtension, file);
-            if (imageReader == null) {
-                LogUtils.info("SVGImagePanel imageReader == null");
-                this.showErrorLabel(fileExtension);
+            SVGImageReader reader = (SVGImageReader) ImageUtils.getTwelveMonkeysRead(svgFile);
+            if (reader == null) {
+                showErrorInfo();
                 return;
             }
-            BufferedImage image = ImageUtils.loadImage(imageReader, 0);
-            ImageUtils.dispose(imageReader);
-            if (image == null) {
-                LogUtils.info("SVGImagePanel bufferedImage == null");
-                this.showErrorLabel(fileExtension);
-                return;
+            try {
+                SVGReadParam param = new SVGReadParam();
+                int displaySize = (int) (sDefaultSize * scale);
+                param.setSourceRenderSize(new Dimension(displaySize, displaySize));
+                BufferedImage image = ImageUtils.loadImage(reader, 0, param);
+                if (image == null) {
+                    showErrorInfo();
+                } else {
+                    int width = reader.getWidth(0);
+                    int height = reader.getHeight(0);
+                    showImageInfo(width, height, image);
+                }
+            } catch (Exception e) {
+                showErrorInfo();
+            } finally {
+                ImageUtils.dispose(reader);
             }
-            showImageInfoLabel(this.bufferedImage = image, file, fileExtension);
         });
-    }
-
-    private void showImageInfoLabel(@NotNull BufferedImage bufferedImage,
-                                    @NotNull VirtualFile file,
-                                    @NotNull String fileExtension) {
-        final String ImageInfo = ImageUtils.getImageInfo(bufferedImage, file, fileExtension);
-        Application application = ApplicationManager.getApplication();
-        application.invokeLater(() -> {
-            JBBox container = JBBox.createHorizontalBox();
-            container.setBorder(BorderFactory.createEmptyBorder(10, 10, 10, 10));
-            container.add(JBBox.createHorizontalGlue());
-            JLabel imageInfoLabel = new JLabel();
-            imageInfoLabel.setText(ImageInfo);
-            container.add(imageInfoLabel);
-            add(container, BorderLayout.NORTH);
-        });
-    }
-
-    private void showErrorLabel(@NotNull String fileExtension) {
-        Application application = ApplicationManager.getApplication();
-        application.invokeLater(() -> {
-            String errorText = String.format("Fail to load %s Image", fileExtension);
-            JLabel errorLabel = new JLabel(errorText, Messages.getErrorIcon(), SwingConstants.CENTER);
-            add(errorLabel, BorderLayout.CENTER);
-        });
-    }
-
-
-    public void registerKeyBoardEvent() {
-        int maskKey = InputEvent.SHIFT_DOWN_MASK;
-        registerKeyboardAction(e -> {
-            this.curScale = curScale + 0.1f;  //放大
-            repaint();
-        }, KeyStroke.getKeyStroke(KeyEvent.VK_W, maskKey), JComponent.WHEN_IN_FOCUSED_WINDOW);
-
-        registerKeyboardAction(e -> {
-            this.curScale = curScale - 0.1f;//缩小
-            repaint();
-        }, KeyStroke.getKeyStroke(KeyEvent.VK_S, maskKey), JComponent.WHEN_IN_FOCUSED_WINDOW);
     }
 
     @Override
     public void paint(Graphics g) {
         super.paint(g);
-        if (bufferedImage == null) return;
-        int imageWidth = bufferedImage.getWidth(null);
-        int imageHeight = bufferedImage.getHeight(null);
-        int displayWidth = (int) (imageWidth * curScale);
-        int displayHeight = (int) (imageHeight * curScale);
-        int displayX = (getWidth() - displayWidth) / 2;
-        int displayY = (getHeight() - displayHeight) / 2;
-        Rectangle srcRect = new Rectangle(0, 0, imageWidth, imageHeight);
-        Rectangle dstRect = new Rectangle(displayX, displayY, displayWidth, displayHeight);
-        StartupUiUtil.drawImage(g, bufferedImage, dstRect, srcRect, null);
+        if (image == null) {
+            g.clearRect(0, 0, getWidth(), getHeight());
+            return;
+        }
+        Rectangle srcRect = new Rectangle(
+                0, 0,
+                image.getWidth(),
+                image.getHeight()
+        );
+        Rectangle dstRect = new Rectangle(
+                (getWidth() - image.getWidth()) / 2,
+                (getHeight() - image.getHeight()) / 2,
+                image.getWidth(),
+                image.getHeight()
+        );
+        StartupUiUtil.drawImage(g, image, dstRect, srcRect, null);
     }
 
     public void dispose() {
-        if (bufferedImage != null) {
-            bufferedImage.flush();
-            bufferedImage = null;
+        if (image != null) {
+            image.flush();
+            image = null;
         }
     }
 }