+++ src/main/resources/html/svga/index.js
@@ -1,46 +1,95 @@
+
 function onPageLoad() {
+    let scale = 1.0
+    let canvasSize
+    const canvasStyle = document.getElementById('canvas').style
+
+    function resizeCanvas() {
+        canvasStyle.width = (canvasSize[0] * scale) + 'px'
+        canvasStyle.height = (canvasSize[1] * scale) + 'px'
+    }
+
+    function calcInitSize(videoSize) {
+        const maxSize = 360
+        const width = videoSize.width
+        const height = videoSize.height
+        if (width < maxSize && height < maxSize) return [width, height];
+        if (height === 0) return [0, 0];
+        let rate = width / height;
+        return width > height ? [maxSize, maxSize / rate] : [maxSize * rate, maxSize];
+    }
+
+    function bindScaleEvent() {
+        function scaleCanvas(delta) {
+            scale += delta
+            if(scale < 0.1) scale = 0.1
+            if(scale > 10) scale = 10
+            resizeCanvas()
+        }
+        window.addEventListener('wheel', function(event) {
+            if (event.ctrlKey || event.metaKey) {
+               event.preventDefault()
+               scaleCanvas(event.deltaY > 0 ? -0.1 : 0.1)
+            }
+        }, { passive: false });
+    }
+
     const player = new SVGA.Player('#canvas')
     const parser = new SVGA.Parser('#canvas')
     parser.load('{file_content_placeholder}', function (videoItem) {
-        fillCanvasInfo(videoItem)
-        resizeCanvas(videoItem)
-        player.setVideoItem(videoItem);
-        player.startAnimation()
+            canvasSize = calcInitSize(videoItem.videoSize)
+            resizeCanvas()
+            fillCanvasInfo(videoItem)
+            bindScaleEvent()
+            player.setVideoItem(videoItem);
+            player.startAnimation()
     })
 }
 
-function resizeCanvas(videoItem) {
-    let videoSize = videoItem.videoSize
-    let size = getFitSize(videoSize.width, videoSize.height)
-    let canvasStyle = document.getElementById('canvas').style
-    canvasStyle.width = size[0].toString().concat('px')
-    canvasStyle.height = size[1].toString().concat('px')
-}
-
-const MAX_SIZE = 320
-
-function getFitSize(width, height) {
-    if (width < MAX_SIZE && height < MAX_SIZE) return [width, height];
-    if (height === 0) return [0, 0];
-    let rate = width / height;
-    return width > height ? [MAX_SIZE, MAX_SIZE / rate] : [MAX_SIZE * rate, MAX_SIZE];
-}
-
 function fillCanvasInfo(videoItem) {
-    let versionDiv = document.getElementById('version_info') //version
+    function getImageInfo(videoItem) {
+        let size = 0;
+        let infoList = []
+        for (let key in videoItem.images) {
+            if (videoItem.images.hasOwnProperty(key)) {
+                let n = getImageSize(videoItem.images[key]);
+                size += n.width * n.height * 4;
+                infoList.push(''.concat(key, ' --- ', "{width: ", n.width.toString(), ", height: ", n.height.toString(), '}'))
+            }
+        }
+        return [convertSize(size), infoList];
+    }
+    function getImageSize(image) {
+        let dec = window.atob(image)
+        let length = dec.length
+        let array = new Uint8Array(new ArrayBuffer(length));
+        let i;
+        for (i = 0; i < length; i++) array[i] = dec.charCodeAt(i);
+        return {width: 256 * array[18] + array[19], height: 256 * array[22] + array[23]}
+    }
+    function convertSize(size) {
+        if (size < 1024) {
+            return size + 'B';
+        } else if (size < 1024 * 1024) {
+            return Math.round(size * 1.0 / 1024) + 'K';
+        } else {
+            return Math.round(size * 1.0 / 1024 / 1024) + 'M';
+        }
+    }
+    const versionDiv = document.getElementById('version_info') //version
     versionDiv.innerText = 'Version: '.concat(videoItem.version);
-    let fpsDiv = document.getElementById('fps_info') //fps
+    const fpsDiv = document.getElementById('fps_info') //fps
     fpsDiv.innerText = 'FPS: '.concat(videoItem.FPS);
-    let frameDiv = document.getElementById('frame_info') //frame
+    const frameDiv = document.getElementById('frame_info') //frame
     frameDiv.innerText = 'Frames: '.concat(videoItem.frames);
-    let sizeDiv = document.getElementById('size_info') //size
+    const sizeDiv = document.getElementById('size_info') //size
     let videoSize = videoItem.videoSize
     sizeDiv.innerText = 'Size: '.concat(videoSize.width, ' x ', videoSize.height);
     let imageInfo = getImageInfo(videoItem) //memory
-    let memoryDiv = document.getElementById('memory_info')
+    const memoryDiv = document.getElementById('memory_info')
     memoryDiv.innerText = 'Memory: '.concat(imageInfo[0]);
     //images
-    let imageListDiv = document.getElementById('image-info-list')
+    const imageListDiv = document.getElementById('image-info-list')
     let innerHTML = ''
     for (let key in imageInfo[1]) {
         if (imageInfo[1].hasOwnProperty(key)) {
@@ -51,39 +100,7 @@ function fillCanvasInfo(videoItem) {
     imageListDiv.innerHTML = innerHTML
 }
 
-function getImageInfo(videoItem) {
-    let size = 0;
-    let infoList = []
-    for (let key in videoItem.images) {
-        if (videoItem.images.hasOwnProperty(key)) {
-            let n = getImageSize(videoItem.images[key]);
-            size += n.width * n.height * 4;
-            infoList.push(''.concat(key, ' --- ', "{width: ", n.width.toString(), ", height: ", n.height.toString(), '}'))
-        }
-    }
-    return [convertSize(size), infoList];
-}
-
-function getImageSize(image) {
-    let dec = window.atob(image)
-    let length = dec.length
-    let array = new Uint8Array(new ArrayBuffer(length));
-    let i;
-    for (i = 0; i < length; i++) array[i] = dec.charCodeAt(i);
-    return {width: 256 * array[18] + array[19], height: 256 * array[22] + array[23]}
-}
-
-function convertSize(size) {
-    if (size < 1024) {
-        return size + 'B';
-    } else if (size < 1024 * 1024) {
-        return Math.round(size * 1.0 / 1024) + 'K';
-    } else {
-        return Math.round(size * 1.0 / 1024 / 1024) + 'M';
-    }
-}
-
 function switchColor(target) {
-    let canvasStyle = document.getElementById('canvas').style
+    const canvasStyle = document.getElementById('canvas').style
     canvasStyle.backgroundColor = getComputedStyle(target, null).backgroundColor
 }